anda abrir incidente

+++ DominioSistema/src/main/java/persistence/DAO/Dao.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface Dao<T> {
     List<T> obtenerTodos();
-    T buscarPorId(int id);
+    T buscarPorId(Long id);
     void agregar(Object unObjeto);
     void modificar(Object unObjeto);
     void eliminar(Object unObjeto);
